+++ library/src/jvmMain/kotlin/dev/nesk/akkurate/constraints/ConstraintViolationSet.kt
@@ -22,14 +22,7 @@ import dev.nesk.akkurate.Path
 public class ConstraintViolationSet internal constructor(private val messages: Set<ConstraintViolation>) : Set<ConstraintViolation> by messages {
     public val byPath: Map<Path, Set<ConstraintViolation>> by lazy { messages.groupBy { it.path }.mapValues { it.value.toSet() } }
 
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as ConstraintViolationSet
-
-        return messages == other.messages
-    }
+    override fun equals(other: Any?): Boolean = messages == other
 
     override fun hashCode(): Int = messages.hashCode()